update amplify model
